requirements.txt
ftfy
regex
tqdm
torch
torchvision
torchaudio
librosa
pillow
pandas
unidecode
phonemizer
inflect
transformers
tensorboard
Cython
matplotlib

opensmile

pyworld
tgt
loguru

lightning
wandb

openai-whisper
jiwer
Resemblyzer
mel-cepstral-distance